Add PixDiaTest to put Dia related PixTests
New class PixDiaTest is created for all TEST_METHOD dependent on dia.
PixTestUtils.cpp/.h is created to share code between PixDiaTest and PixTest.

Will enable PixTest for linux next.

This is for https://github.com/microsoft/DirectXShaderCompiler/issues/5877
