+++ tools/clang/unittests/HLSL/CMakeLists.txt
@@ -42,7 +42,9 @@ add_clang_library(ClangHLSLTests SHARED
   Objects.cpp
   OptimizerTest.cpp
   OptionsTest.cpp
+  PixDiaTest.cpp
   PixTest.cpp
+  PixTestUtils.cpp
   RewriterTest.cpp
   SystemValueTest.cpp
   ValidationTest.cpp
@@ -53,6 +55,7 @@ add_clang_library(ClangHLSLTests SHARED
 else (WIN32)
 set(HLSL_IGNORE_SOURCES
   MSFileSysTest.cpp
+  PixDiaTest.cpp
   PixTest.cpp
   RewriterTest.cpp
   )
@@ -71,6 +74,7 @@ add_clang_unittest(ClangHLSLTests
   Objects.cpp
   OptimizerTest.cpp
   OptionsTest.cpp
+  PixTestUtils.cpp
   SystemValueTest.cpp
   TestMain.cpp
   ValidationTest.cpp
